4896 - Keep original order when selecting Measures/Dimensions
